Conflicts:
	src/corelib/io/io.pri
	src/corelib/thread/qwaitcondition_unix.cpp
	src/network/access/qnetworkaccessmanager.cpp
	src/plugins/plugins.pro

+++ src/corelib/kernel/qeventdispatcher_win_p.h
@@ -58,7 +58,7 @@ class QEventDispatcherWin32Private;
 
 // forward declaration
 LRESULT QT_WIN_CALLBACK qt_internal_proc(HWND hwnd, UINT message, WPARAM wp, LPARAM lp);
-int qt_msectime();
+quint64 qt_msectime();
 
 class Q_CORE_EXPORT QEventDispatcherWin32 : public QAbstractEventDispatcher
 {
