+++ src/gui/painting/qpainter_p.h
@@ -240,7 +240,7 @@ public:
     }
 
     QTransform viewTransform() const;
-    int effectiveDevicePixelRatio() const;
+    qreal effectiveDevicePixelRatio() const;
     QTransform hidpiScaleTransform() const;
     static bool attachPainterPrivate(QPainter *q, QPaintDevice *pdev);
     void detachPainterPrivate(QPainter *q);
